mna-extractor/pyproject.toml
[project]
name = "mna-extractor"
version = "0.1.0"
description = "LLM-based extraction pipeline for M&A buyer firm websites; outputs 33-column Buyer Database xlsx with [REQUIRES EXTERNAL] flagging for PE-internal numerical fields"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "openai>=1.40",
    "tenacity>=8.5",
    "tldextract>=5.1",
    "selectolax>=0.3.21",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "openpyxl>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]
playwright = [
    "playwright>=1.45",
]

[project.scripts]
mna-extract = "mna_extractor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mna_extractor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
